% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspl.R
\name{bspl}
\alias{bspl}
\title{B-spline fits on Multivariate Hydrographic Profiles}
\usage{
bspl(Pi, Xi, nbas = 20, fdn = list("Temperature", "Salinity"))
}
\arguments{
\item{Pi}{vector containing the levels}

\item{Xi}{array containing the profiles stored in this order \code{levels} x \code{stations} x \code{variables}}

\item{nbas}{number of Bsplines (coefficients), by default nbas = 20.}

\item{fdn}{a list of the variable names, by default fdn = list('Temperature','Salinity').}
}
\value{
\code{fdobj} : fd objects of the splines construction containing coefficients, basis etc... The coefficients are stored in an array \code{nbasis} x \code{stations} x \code{variables}
}
\description{
This function fits B-splines on multivariate hydrographic profiles and return a functional data object.
}
\references{
Pauthenet et al. (2017) A linear decomposition of the Southern Ocean thermohaline structure. Journal of Physical Oceanography, http://dx.doi.org/10.1175/JPO-D-16-0083.1

Ramsay, J. O., and B. W. Silverman, 2005: Functional Data Analysis. 2nd Edition Springer, 426 pp., Isbn : 038740080X.
}
\seealso{
\code{\link{fpca}} for functional principal component analysis of T-S profiles, \code{\link{proj}} for computing Principal Components, \code{\link{reco}} for reconstructing profiles with less modes.
}
\author{
Etienne Pauthenet \email{<etienne.pauthenet@gmail.com>}, David Nerini \code{<david.nerini@univ-amu.fr>}, Fabien Roquet \code{<fabien.roquet@gu.se>}
}
