% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenf.R
\name{eigenf}
\alias{eigenf}
\title{Plot of the vertical modes}
\usage{
eigenf(pca, te, le = 50, sign = 1)
}
\arguments{
\item{pca}{the list produced by the function \code{fpca}.}

\item{te}{Which PC to plot.}

\item{le}{number of '+' or '-' by curve, default at \code{le = 50}}

\item{sign}{The PCs are invariant by their sign, sign can contain a vector of \code{1} or \code{-1} to inverse the sign of the PCs if wanted. \code{sign} can also be used as a factor to increase the effect of eigenfunctions and see better the small variations.}
}
\value{
Representation of the effects of the vertical mode \code{te} on temperature and salinity mean profiles. The curves show the mean profile (solid) and the effects of adding (+) and subtracting (-) eigenfunctions. The percentages next to the header titles are the amount of variance explained by the mode displayed. The percentages in the horizontal axis label are the variance contained by each variable (T and S) on the mode displayed. \code{sign} can be used as a factor to enhance the deformation of + and - curves.
}
\description{
Plot the effect of each vertical mode (i. e. eigenfunction) on the mean Temperature and Salinity (T-S) profile.
}
\references{
Ramsay J. O., and B. W. Silverman, 2005: Functional Data Analysis. Springer, 426 pp.
}
\seealso{
\code{\link{bspl}} for bsplines fit on T-S profiles, \code{\link{fpca}} for functional principal component analysis of T-S profiles, \code{\link{PCmap}} for plotting a map of PC
}
