% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca.R
\name{fpca}
\alias{fpca}
\title{Functional Principal Component Analysis (FPCA) of T-S profiles}
\usage{
fpca(temp.fd, sal.fd, plot = FALSE, plot3d = FALSE)
}
\arguments{
\item{temp.fd, sal.fd}{fd objects (list) of the splines construction containing coefficients, etc... This is produced by the function \code{bspl}.}

\item{plot, plot3d}{if TRUE, plot the first two or three PCs.}
}
\value{
\code{pca} list containing the eigen (\code{values}), eigen (\code{vectors}), eigen vectors not weighted by WM (\code{vecnotWM}), principal components (\code{pc}), deformation induced by an axis (\code{axe}) percentage of each axis (\code{pval}).
}
\description{
FPCA in the multivariate case, applied on T-S profiles seen as curves (Bsplines)
}
\references{
Ramsay J. O., and B. W. Silverman, 2005: Functional Data Analysis. Springer, 426 pp.
}
\seealso{
\code{\link{bspl}} for bsplines fit on T-S profiles, \code{\link{PCmap}} for plotting a map of PC, \code{\link{kde_pc}} for kernel density estimation of two PCs...
}
