% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_pc.R
\name{kde_pc}
\alias{kde_pc}
\title{Kernel Density Estimation of two Principal Components}
\usage{
kde_pc(pca, te = c(1, 2), h = 0.2, n = 50, l = seq(0.01, 4, 0.3),
  title = "", c = "grey15")
}
\arguments{
\item{pca}{The \code{pca} list returned by the function \code{fpca}.}

\item{te}{A vector of which two PCs to plot against each other. The default values are PC1/PC2 (i.e. te = c(1,2)) makes a lot of sense when looking at T-S profiles.}

\item{h}{vector of bandwidths for x and y directions. Defaults is h = 0.2.}

\item{n}{Number of grid points in each direction. Can be scalar or a length-2 integer vector. Default is h = 50.}

\item{l}{levels of contour, default is l = seq(.1,4,.2).}

\item{c}{Color of the colorbar, default is grey.}
}
\value{
A kernel density plot of two PCs i.e. a 2D boxplot.
}
\description{
Produce a Kernel Density Estimation (KDE) plot of two Principal Components (PCs) (using \code{kde2d {MASS}}), on the axis are the percentage of explained variance of each PC (pca$pval).
}
\details{
This function is made to be modified at will, to add contours, change colors etc...
}
\seealso{
\code{\link{bspl}} for bsplines fit on T-S profiles, \code{\link{fpca}} for functional principal component analysis of T-S profiles, \code{\link{PCmap}} for plotting a map of PC
}
