% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_pc.R
\name{map_pc}
\alias{map_pc}
\title{Map of PC}
\usage{
map_pc(lon, lat, pca, mask, te, sign = 1)
}
\arguments{
\item{lon}{A vector of longitude.}

\item{lat}{A vector of latitude.}

\item{pca}{the list produced by the function \code{fpca}.}

\item{te}{Which PC is in the \code{PC} matrix.}

\item{sign}{The PCs are invariant by their sign, sign can contain a vector of \code{1} or \code{-1} to inverse the sign of the PCs if wanted.}
}
\value{
A color map of the PC.
}
\description{
Produce a map of PC, with the colorbar centered in zero, the coastline, the percentage of explained variance associated with the PC (pca$pval).
}
\details{
This function is made to be modified at will, to add contours, change colors etc...
}
\seealso{
\code{\link{bspl}} for bsplines fit on T-S profiles, \code{\link{fpca}} for functional principal component analysis of T-S profiles,...
}
