% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_plot.R
\name{pc_plot}
\alias{pc_plot}
\title{Graphical representation of the PC on the main principal plans}
\usage{
pc_plot(pca, pc, n = c(1, 2))
}
\arguments{
\item{\code{pca}}{list containing the vertical modes used (computed with the function \code{fpca}).}

\item{\code{pc}}{The principal components of the profiles (computed with the function \code{proj}).}

\item{\code{n}}{vector of the PC to plot, it can be of length 2 or 3. Default is n = c(1,2)}
}
\value{
Plot the main PC with their corresponding variance on the axis labels.

#' @author Etienne Pauthenet \email{<etienne.pauthenet@gmail.com>}, David Nerini \code{<david.nerini@univ-amu.fr>}, Fabien Roquet \code{<fabien.roquet@gu.se>}
}
\description{
This function plots the PC values with the appropriate axis and percentage values on the labels.
}
\references{
Pauthenet et al. (2017) A linear decomposition of the Southern Ocean thermohaline structure. Journal of Physical Oceanography, https://doi.org/10.1175/JPO-D-16-0083.1

Ramsay, J. O., and B. W. Silverman, 2005: Functional Data Analysis. 2nd Edition Springer, 426 pp., Isbn : 038740080X.
}
\seealso{
\code{\link{bspl}} for bsplines fit on T-S profiles, \code{\link{fpca}} for functional principal component analysis of T-S profiles, \code{\link{proj}} for computing Principal Components, \code{\link{reco}} for reconstructing profiles with less modes.
}
