% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspl.R
\name{bspl}
\alias{bspl}
\title{Fit of B-splines on Temperature and Salinity profiles}
\usage{
bspl(temp, sal, Depth, range, mybn = 20, l = 100)
}
\arguments{
\item{temp}{temperature stored in arrays \code{ind*depth} or \code{lon*lat*depth} or \code{lon*lat*depth*time}}

\item{sal}{salinity stored in arrays \code{lon*lat*depth} or \code{lon*lat*depth*time}}

\item{range}{vector of two elements containing the first and last depth level \code{c(dmin,dmax)}, \code{dmin} and \code{dmax} must be contained by the \code{depth} range.}

\item{mybn}{number of splines, Default is mybn = 20.}

\item{l}{number of interpolation point. The number of points to interpolate must be large enough to avoid overfitting. Default is l = 100.}

\item{depth}{depth can be a vector or a matrix (case of irregular vertical sampling)}
}
\value{
\code{temp.fd} and \code{sal.fd} : fd objects of the splines construction containing coefficients, etc...

\code{BS} list containing :\code{profx}, \code{Teval} and \code{Seval} (the splines sampled along profx)

\code{mask} : if the data was on a regular horizontal grid
}
\description{
Fit B-splines on Temperature and Salinity (T-S) profiles from gridded dataset (reanalysis, models,...)
}
\seealso{
\code{\link{fpca}} for functional principal component analysis of T-S profiles, \code{\link{PCmap}} for plotting a map of PC, \code{\link{kde_pc}} for kernel density estimation of two PCs...
}
