% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda.oce.R
\docType{package}
\name{fda.oce}
\alias{fda.oce}
\alias{fda.oce-package}
\title{fda.oce: A package for the use of functional data analysis tools on oceanographic profiles.}
\description{
The fda.oce package provides two categories of functions : computation :
\code{bspl}, \code{fpca}, \code{reco}, \code{proj}
and visualisation :
\code{eigenf},\code{map_pc},\code{kde_pc}.
}
