% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{reco}
\alias{reco}
\title{Reconstruction of T-S profiles}
\usage{
reco(pca, te)
}
\arguments{
\item{pca}{list produced by the function \code{fpca}}

\item{te}{how many PCs to use in the reconstruction, default is set to the total number of PC, \code{te = mybn}.}
}
\value{
\code{recotemp} and \code{recosali} matrix containing the reconstruction of \code{temp} and \code{sali} with the numer of PCs \code{te}.
}
\description{
Reconstruction of T-S profiles with a chosen number of Principal Components (PCs).
}
\references{
Ramsay J. O., and B. W. Silverman, 2005: Functional Data Analysis. Springer, 426 pp.
}
\seealso{
\code{\link{bspl}} for bsplines fit on T-S profiles, \code{\link{fpca}} FPCA of T-S profiles, \code{\link{PCmap}} for plotting a map of PC, \code{\link{kde_pc}} for kernel density estimation of two PCs...
}
